Introduce a new <type> attribute called luks_pbkdf
This attribute can be used to control what algorithm LUKS will use
to derive a symmetric key from a password when unlocking a key slot.

The default behavior is to use whatever cryptsetup picks as the default.
To override that choice, specify luks_pbkdf="ALGO" where ALGO can be one
of pbkdf2, argon2i, or argon2id, respectively.

Signed-off-by: Olaf Kirch <okir@suse.com>