Fixed image install of dnf package manager
If the package manager is set to 'dnf4', this name is
taken also as package name to install this package manager
into the image. Problem is 'dnf4' is not a valid package
name. The correct name for 'dnf4' is 'dnf'. This commit
fixes the name bug introduced with the changes for
Issue #2262