Add alternative re-reading of the partition table
To inform the kernel about disk geometry changes, kiwi uses
partprobe as a primary tool. However it is provided by parted
and not necessarily available due to the package requirements
on the dracut module. A second attempt via blockdev which
is expected to exist is therefore made by this commit