Change location of LUKS keyfile to /root/.root.keyfile
In some configuration (eg on a transactional system), the root
file system is read-only, which means we will not be able to remove
the stale .root.keyfile after firstboot.

This patch changes the location of the keyfile and places it in
/root/.root.keyfile instead, because root's home directory is much
more likely to be read/write.

Also fixed the associated unit test (thanks to schaefi).

Signed-off-by: Olaf Kirch <okir@suse.com>