Improve loopback.cfg example
submenu+source should be preferred over menuentry+configfile:

- submenu creates a new context, menuentry does not. As a result, changes made
to vars inside menuentry's braces pollute the current scope. In this particular
case what is notably wrong is root=loop happens in the current scope. If we
activate the "Boot from openSUSE ISO" menu entry and then press Esc to return
back to explore other menu entries, $root would still be set to 'loop'. Instead
of manually saving and restoring $root value it is better to wrap the whole
thing in a new context with submenu

- configfile is not just 'source in a new context', it has as least one side
effect. For details look for grub_err_printed_errors usage in [1] and [2] in
grub sources

[1] grub-core/normal/menu.c
[2] grub-core/normal/menu_entry.c