Use bootpartition for Fedora integration test
When using btrfs with the proposed layout for testing the
delivered grub bios module for the Fedora system used to build
the integration test (FC37) is not capable to find the grub
config file. A manual call for configfile in the grub shell
fixes this with the existing kiwi created grub early-boot
script. However, it is expected that the delivered grub image
works and kiwi only creates its own one if no distro delivered
grub image was found. To make the integration test functional
for both BIOS and EFI the simple solution is to use an extra
not btrfs based boot partition. This still allows to test
the desired btrfs layout in terms of volumes and sub-volumes
and does not break on any of the boot methods.