Switch to pyproject.toml
Calling setup.py is marked obsolete for a long time and with this commit
pyproject.toml is introduced and used for the development environment
setup. On the packaging side the spec file is kept compatible with
setup.py in a way that we try the build and installer python modules
first and only if those are not present setup.py is called. For the
publishing on pypi both targets wheel and sdist are used which is the
default action of python -m build.