Replace the regex-based loader entry fix with string parsing (#2388)
A user building RHEL images ran into issues with the initrd. 
It turns out that RHEL uses some patches that mean the
initrd/linux files in RHEL are not installed to /boot, which trips
up the original regex. The new fix doesn't rely on matching the
path in boot, instead just finding the initrd/linux files and rewriting
them in place.

This change also adds the pre-and-post fix loader entries to the debug logs.

Reference: https://bugzilla.suse.com/1208701

Fixes suse bsc#1208701