Ensure setfiles is detected inside the image-root
We do not actually use setfiles from the host, we use it from the
image root we create for the image build. Thus, we should look in
the image root instead of on the host system.

This prevents us from incorrectly detecting that setfiles is not
available for setting SELinux contexts.

Fixes: https://github.com/OSInside/kiwi/issues/2414
Fixes: 2a22901ddd11ae23b6724b5e1aaa4261f219ccb6