Allow newer versions of sphinx again
With Sphinx >= 6.0.0, jQuery is no longer included but the rtd theme's search
functionality needs it. The theme was changed to load sphinxcontrib-jquery
automatically with https://github.com/readthedocs/sphinx_rtd_theme/pull/1399,
but kiwi broke this fix by loading the theme in the wrong way.

Now that the theme is loaded properly, the search functionality also works
with newer versions of sphinx again and the version pinning can be lifted.

Fixes: #2462