Don't use LUKS keyfile with disk password
In case a disk password is configured, no keyfile
will be used. The setup of the disk password is a
method to store an insecure password into the bootloader
configuration and directly call cryptomount with
that password to boot into the system once. The
insecure credentials will then be replaced by a
secure version e.g from a TPM and a final keyfile
will be placed into the luks pool. Thus there is
no need for kiwi to create an initial keyfile as
it's insecure and overwritten anyway. This is
related to bsc#1218181