Refactor Command.run() to return None if cmd not found
Command.run() currently has a bit of a confusing behavior: if raise_on_error is
False and the executable is not found, then a weird CommandT is returned (return
code is -1 and stdout+stderr is None). This makes it possible to hanlde command
not found errors separately, but it makes that needlessly verbose. So instead,
let's just return None in *this* special case.

That in turn uncovered, that in most cases when we set `raise_on_error=True`, we
actually want an error if the command is not present but no error if the command
fails to execute (e.g. because it returns -1 if you run `$cmd --version`
🙄). Hence we introduce the flag `raise_on_command_not_found`, which causes an
exception to be raised if the command is not found. This makes it independent of
the `raise_on_error` flag.

Additionally, we add a small optimization: if command starts with /, then we
assume it's a full path and we omit the call to which (and just check whether it
exists).

These changes require additional tests and a few assertions in places where mypy
cannot infer that a certain value is of the correct type.