dracut resizing: on GPT don't assume resize
When the script in the dracut module checks if a resize should take
place, it checks if the partition table was expanded to cover the
entirety of the disk. If the check says that this has happened,
partitions are not resized either. This check only exists for GPT.

There is hardware with UEFI firmware that on reboot "fixes up" the GPT table
and expands it to the entirety of the disk. This causes partition
expansion to be skipped.