kiwi no longer uses debootstrap
For building Debian based images we used debootstrap to
bootstrap an empty root until apt-get could be used to
complete the job. This has now changed such hat apt-get
is also used for bootstrapping a new system. The concept
and also potential alternatives to the way kiwi bootstraps
Debian based systems can be found here:

* https://osinside.github.io/kiwi/working_with_images/build_without_debianbootstrap.html

Due to the drop of debootstrap it might happen that
package lists of existing image descriptions needs to be
extended with packages that were formerly pulled in by
debootstrap but did not get properly pulled in with the
new apt based bootstrap. As reference please check out the
integration tests from here:

* https://github.com/OSInside/kiwi/tree/main/build-tests/x86/ubuntu
* https://github.com/OSInside/kiwi/tree/main/build-tests/x86/debian

Thanks