Hackweek: Add TargetDevice class
Allow to build the image directly to a specified target
block device. The idea here is that this feature could
be used for an installer-to-target. For example an
installer ISO image that holds a distribution can use
the kiwi API to build and deploy a system on target.
With this concept also some of the initrd based system
modifications via ignition/combustion are no longer
required