Fix globbing with exclude with regex
This fixes a collection of bugs when producing erofs images.

On one hand, this ensures that an exclude of `/sys` doesn't accidentally
match `/lib/libsystemd.so`, only `/sys/whatever`.

On the other hand, this ensures that `/dev/*` does match `/dev/vda` and
not just `/dev///////////`.

This fixes libsystemd.so getting dropped in Kiwi-built FEX images.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>