package_manager: apt: run postinst scripts for ca-certificates
The ca-certificates package is required when any of the repositories
being used for the image use HTTPS. In order to support this, the
package needs to be added to the bootstrap packages. However, the
bootstrap package installer will not run the postinst script, which is
required to configure the system to use the SSL certificates.

This patch adjusts the bootstrap package installer to additionally run
the postinst scripts for ca-certificates.

Signed-off-by: Isaac True <isaac@is.having.coffee>