Added OCI container app integration test
Builds a container that only contains the application and needed
tools and does not provide other data. This integration test
serves as an example how to create a small container without
core components like a shell or coreutils, unneeded libraries, etc..
As example app the flake-pilot package is used which is a rust
based application.