Run %check for TW only
For full test coverage the anymarkup module must be installed.
However the module is an optional module for kiwi and only
required when running tests to reach the 100% code coverage.
Unfortunately not all distros provide anymarkup which is the
reason why I limited the %check execution to only TW for the
moment.