Make unit tests to pass
Added a small refactor to prevent code duplication and applied
tests for code parts that I believe are good to go. For other
code parts I added a "TODO: ostree:" marker and spared out the
code from being covered by the coverage checker. For those
remaining changes we need to have a conversation how to proceed