Use GPT table for EFI capable ISO images
If the ISO should be able to boot via EFI make sure the
partition table is GPT. In addition to that setting also
make sure the CSM boot flag for the embedded MBR table
is provided through an extra bearer partition in the same
way as it's done for EFI capable disk images. For better
debugging also add the report of the system area as
debug information to the log file. This Fixes #2685