Update system files setup for containers
The attribute provide_system_files creates a meta file in the
root tree named 'systemfiles'. The contents of this file were
produced by just a dump of the package database so far. For
a more generic use of this data some adaptions were needed.
First we allow to skip packages matching a pattern from being
part of the system files. Next we do not put ghost and doc
files into account. And last we do not add symlink targets
if the target path is also part of the package. The consumer
of this information is flake-pilot which syncs the system files
from the host via --copy-links. This allows a more generic
use with regards to versioned libraries e.g. libc